requirements.txt
fastapi
uvicorn
google-genai
neo4j
python-dotenv
